+++ collectors/bzimport/tests/test_fixups.py
@@ -1,6 +1,7 @@
 import pytest
 
-from collectors.bzimport.fixups import AffectFixer
+from collectors.bzimport.fixups import AffectFixer, FlawFixer
+from osidb.models import Flaw
 
 pytestmark = pytest.mark.unit
 
@@ -19,3 +20,96 @@ class TestAffectFixer:
 
         # should not change the affect with correct module at all
         assert correct_module == AffectFixer.fixplace_ps_module(correct_module)
+
+
+class TestFlawFixer:
+    @pytest.mark.parametrize(
+        "summary,title,component",
+        [
+            (
+                "carbon: cheesecake",
+                "cheesecake",
+                "carbon",
+            ),
+            (
+                "EMBARGOED carbon: cheesecake",
+                "cheesecake",
+                "carbon",
+            ),
+            (
+                "CVE-2000-12345 CVE-3000-12345 carbon: cheesecake",
+                "cheesecake",
+                "carbon",
+            ),
+            (
+                "EMBARGOED CVE-3000-12345 carbon: cheesecake",
+                "cheesecake",
+                "carbon",
+            ),
+            (
+                "radioactive carbon: cheesecake",
+                "cheesecake",
+                "carbon",
+            ),
+            (
+                "radioactive carbon : cheesecake",
+                "radioactive carbon : cheesecake",
+                "",
+            ),
+            (
+                "carbon cheesecake:",
+                "",
+                "cheesecake",
+            ),
+            (
+                "carbon cheesecake",
+                "carbon cheesecake",
+                "",
+            ),
+            (
+                "EMBARGOED",
+                "",
+                "",
+            ),
+            (
+                "CVE-2000-12345 CVE-3000-12345",
+                "",
+                "",
+            ),
+            (
+                "EMBARGOED CVE-2000-12345 CVE-3000-12345",
+                "",
+                "",
+            ),
+            (
+                "EMBARGOED:",
+                ":",
+                "",
+            ),
+            (
+                "radioactive: carbon: cheesecake:",
+                "",
+                "cheesecake",
+            ),
+            (
+                "radioactive: carbon: cheesecake",
+                "cheesecake",
+                "carbon",
+            ),
+            (
+                "   carbon:   cheesecake   ",
+                "cheesecake",
+                "carbon",
+            ),
+        ],
+    )
+    def test_fix_title(self, summary, title, component):
+        """
+        test title fixup which performs mapping from Bugzilla summary
+        it not only sets the correct flaw title but also the component
+        """
+        flaw = Flaw()
+        flaw_fixer = FlawFixer(flaw, {"summary": summary}, None)
+        flaw_fixer.fix_title()
+        assert flaw.title == title
+        assert flaw.component == component