+++ osidb/models.py
@@ -698,24 +698,6 @@ class Flaw(
         # users from reviewing in the first place, in which case we don't
         # need to perform this validation
 
-    def _validate_public_flaw_title(self):
-        """
-        Check that the flaw's title does not contain the word "EMBARGOED" if the flaw is public.
-        """
-        if not self.is_embargoed and "EMBARGOED" in self.title:
-            raise ValidationError(
-                'Flaw title contains "EMBARGOED" despite being public.'
-            )
-
-    def _validate_embargoed_flaw_title(self):
-        """
-        Check that the flaw's title does contain the word "EMBARGOED" if the flaw is embargoed.
-        """
-        if self.is_embargoed and "EMBARGOED" not in self.title:
-            raise ValidationError(
-                'Flaw title does not contain "EMBARGOED" despite being embargoed.'
-            )
-
     def _validate_embargoing_public_flaw(self):
         """
         Check whether a currently public flaw is being embargoed.