+++ osidb/tests/test_flaw.py
@@ -906,41 +906,6 @@ class TestFlawValidators:
         else:
             assert flaw1.save() is None
 
-    @pytest.mark.parametrize(
-        "title,should_raise",
-        [
-            ("CVE-2022-1234 kernel: some description", False),
-            ("EMBARGOED CVE-2022-1234 kernel: some description", True),
-        ],
-    )
-    def test_validate_public_flaw_title(self, title, should_raise):
-        """test that public flaws only accepts a valid title without the "EMBARGOED" word"""
-        if should_raise:
-            with pytest.raises(ValidationError) as e:
-                FlawFactory(embargoed=False, title=title)
-            assert 'Flaw title contains "EMBARGOED" despite being public.' in str(e)
-        else:
-            assert FlawFactory(embargoed=False, title=title)
-
-    @pytest.mark.parametrize(
-        "title,should_raise",
-        [
-            ("EMBARGOED CVE-2022-1234 kernel: some description", False),
-            ("CVE-2022-1234 kernel: some description", True),
-        ],
-    )
-    def test_validate_embargoed_flaw_title(self, title, should_raise):
-        """test that embargoed flaws only accepts a valid title containing the "EMBARGOED" word"""
-        if should_raise:
-            with pytest.raises(ValidationError) as e:
-                FlawFactory(embargoed=True, title=title)
-            assert (
-                'Flaw title does not contain "EMBARGOED" despite being embargoed.'
-                in str(e)
-            )
-        else:
-            assert FlawFactory(embargoed=True, title=title)
-
     @freeze_time(tzdatetime(2021, 11, 23))
     def test_validate_embargoing_public_flaw(self):
         flaw = FlawFactory(embargoed=False)